Replace libsignal-protocol-c with libomemo-c
The library libsignal-protocol-c is not maintained. This commit will
replace the dependency libsignal-protocol-c with libomemo-c.

On 2025-02-07 a new version of protobuf-c has been uploaded to
Debian. This new version induced a bug for libsignal-protocol-c
and libomemo-c:

 * Bug#1095352: libomemo-c: FTBFS [1]
 * Bug#1095353: libsignal-protocol-c: FTBFS [2]

(Thanks to Dino team and Debian Maintainer it's solved)

libomemo-c [3] is a fork of libsignal-protocol-c [4] maintained
by the Dino team.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1095352
[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1095353
[3] https://github.com/dino/libomemo-c
[4] https://github.com/signalapp/libsignal-protocol-c

Note:
Commits squashed and commit message edited by @jubalh.